+++ docs-util/helpers/use-db.js
@@ -1,26 +1,26 @@
-const { dropDatabase, createDatabase } = require("pg-god");
-const { createConnection } = require("typeorm");
+const { dropDatabase, createDatabase } = require("pg-god")
+const { createConnection } = require("typeorm")
 
-const path = require("path");
+const path = require("path")
 
 const DbTestUtil = {
   db_: null,
 
   setDb: function (connection) {
-    this.db_ = connection;
+    this.db_ = connection
   },
 
   clear: function () {
-    return this.db_.synchronize(true);
+    return this.db_.synchronize(true)
   },
 
   shutdown: async function () {
-    await this.db_.close();
-    return dropDatabase({ databaseName });
+    await this.db_.close()
+    return dropDatabase({ databaseName })
   },
-};
+}
 
-const instance = DbTestUtil;
+const instance = DbTestUtil
 
 module.exports = {
   initDb: async function ({ cwd }) {
@@ -33,19 +33,19 @@ module.exports = {
         `dist`,
         `migrations`
       )
-    );
+    )
 
-    const databaseName = "medusa-fixtures";
-    await createDatabase({ databaseName });
+    const databaseName = "medusa-fixtures"
+    await createDatabase({ databaseName })
 
     const connection = await createConnection({
       type: "postgres",
       url: "postgres://localhost/medusa-fixtures",
       migrations: [`${migrationDir}/*.js`],
-    });
+    })
 
-    await connection.runMigrations();
-    await connection.close();
+    await connection.runMigrations()
+    await connection.close()
 
     const modelsLoader = require(path.join(
       cwd,
@@ -55,19 +55,19 @@ module.exports = {
       `dist`,
       `loaders`,
       `models`
-    )).default;
+    )).default
 
-    const entities = modelsLoader({}, { register: false });
+    const entities = modelsLoader({}, { register: false })
     const dbConnection = await createConnection({
       type: "postgres",
       url: "postgres://localhost/medusa-fixtures",
       entities,
-    });
+    })
 
-    instance.setDb(dbConnection);
-    return dbConnection;
+    instance.setDb(dbConnection)
+    return dbConnection
   },
   useDb: function () {
-    return instance;
+    return instance
   },
-};
+}