+++ integration-tests/api/__tests__/store/cart.js
@@ -6,6 +6,7 @@ const { useApi } = require("../../../helpers/use-api")
 const { initDb, useDb } = require("../../../helpers/use-db")
 
 const cartSeeder = require("../../helpers/cart-seeder")
+const swapSeeder = require("../../helpers/swap-seeder")
 
 jest.setTimeout(30000)
 
@@ -20,8 +21,12 @@ describe("/store/carts", () => {
 
   beforeAll(async () => {
     const cwd = path.resolve(path.join(__dirname, "..", ".."))
-    dbConnection = await initDb({ cwd })
-    medusaProcess = await setupServer({ cwd })
+    try {
+      dbConnection = await initDb({ cwd })
+      medusaProcess = await setupServer({ cwd })
+    } catch (error) {
+      console.log(error)
+    }
   })
 
   afterAll(async () => {
@@ -96,6 +101,7 @@ describe("/store/carts", () => {
     beforeEach(async () => {
       try {
         await cartSeeder(dbConnection)
+        await swapSeeder(dbConnection)
       } catch (err) {
         console.log(err)
         throw err
@@ -104,6 +110,24 @@ describe("/store/carts", () => {
 
     afterEach(async () => {
       await doAfterEach()
+    })
+
+    // We were experiencing some issues when having created a cart in a region
+    // containing multiple countries. At this point, the cart does not have a shipping
+    // address. Therefore, on subsequent requests to update the cart, the server
+    // would throw a 500 due to missing shipping address id on insertion.
+    it("updates a cart, that does not have a shipping address", async () => {
+      const api = useApi()
+
+      const response = await api.post("/store/carts", {
+        region_id: "test-region-multiple",
+      })
+
+      const getRes = await api.post(`/store/carts/${response.data.cart.id}`, {
+        region_id: "test-region",
+      })
+
+      expect(getRes.status).toEqual(200)
     })
 
     it("fails on apply discount if limit has been reached", async () => {
@@ -327,9 +351,82 @@ describe("/store/carts", () => {
         expect(e.response.status).toBe(409)
       }
 
-      //check to see if payment has been cancelled
+      //check to see if payment has been cancelled and cart is not completed
       const res = await api.get(`/store/carts/test-cart-2`)
       expect(res.data.cart.payment.canceled_at).not.toBe(null)
+      expect(res.data.cart.completed_at).toBe(null)
+    })
+
+    it("fails to complete swap cart with items inventory not/partially covered", async () => {
+      const manager = dbConnection.manager
+
+      const li = manager.create(LineItem, {
+        id: "test-item",
+        title: "Line Item",
+        description: "Line Item Desc",
+        thumbnail: "https://test.js/1234",
+        unit_price: 8000,
+        quantity: 99,
+        variant_id: "test-variant-2",
+        cart_id: "swap-cart",
+      })
+      await manager.save(li)
+
+      await manager.query(
+        "UPDATE swap SET cart_id='swap-cart' where id='test-swap'"
+      )
+
+      const api = useApi()
+
+      try {
+        await api.post(`/store/carts/swap-cart/complete-cart`)
+      } catch (e) {
+        expect(e.response.data).toMatchSnapshot({
+          code: "insufficient_inventory",
+        })
+        expect(e.response.status).toBe(409)
+      }
+
+      //check to see if payment has been cancelled and cart is not completed
+      const res = await api.get(`/store/carts/swap-cart`)
+      expect(res.data.cart.payment_authorized_at).toBe(null)
+      expect(res.data.cart.payment.canceled_at).not.toBe(null)
+    })
+
+    it("successfully completes swap cart with items inventory not/partially covered due to backorder flag", async () => {
+      const manager = dbConnection.manager
+
+      const li = manager.create(LineItem, {
+        id: "test-item",
+        title: "Line Item",
+        description: "Line Item Desc",
+        thumbnail: "https://test.js/1234",
+        unit_price: 8000,
+        quantity: 99,
+        variant_id: "test-variant-2",
+        cart_id: "swap-cart",
+      })
+      await manager.save(li)
+      await manager.query(
+        "UPDATE swap SET cart_id='swap-cart' where id='test-swap'"
+      )
+      await manager.query(
+        "UPDATE swap SET allow_backorder=true where id='test-swap'"
+      )
+      await manager.query("DELETE FROM payment where swap_id='test-swap'")
+
+      const api = useApi()
+
+      try {
+        await api.post(`/store/carts/swap-cart/complete-cart`)
+      } catch (error) {
+        console.log(error)
+      }
+
+      //check to see if payment is authorized and cart is completed
+      const res = await api.get(`/store/carts/swap-cart`)
+      expect(res.data.cart.payment_authorized_at).not.toBe(null)
+      expect(res.data.cart.completed_at).not.toBe(null)
     })
   })