+++ packages/medusa/src/api/routes/store/customers/reset-password-token.js
@@ -1,12 +1,10 @@
 import { MedusaError, Validator } from "medusa-core-utils"
 
 /**
- * @oas [post] /customers/{id}/password-token
+ * @oas [post] /customers/password-token
  * operationId: PostCustomersCustomerPasswordToken
  * summary: Creates a reset password token
  * description: "Creates a reset password token to be used in a subsequent /reset-password request. The password token should be sent out of band e.g. via email and will not be returned."
- * parameters:
- *   - (path) id=* {string} The id of the Customer.
  * tags:
  *   - Customer
  * responses:
@@ -15,9 +13,7 @@ import { MedusaError, Validator } from "medusa-core-utils"
  */
 export default async (req, res) => {
   const schema = Validator.object().keys({
-    email: Validator.string()
-      .email()
-      .required(),
+    email: Validator.string().email().required(),
   })
 
   const { value, error } = schema.validate(req.body)