+++ packages/medusa/src/models/product.ts
@@ -26,6 +26,13 @@ import { ProductVariant } from "./product-variant"
 import { ShippingProfile } from "./shipping-profile"
 import _ from "lodash"
 
+export enum Status {
+  DRAFT = "draft",
+  PROPOSED = "proposed",
+  PUBLISHED = "published",
+  REJECTED = "rejected",
+}
+
 @Entity()
 export class Product {
   @PrimaryColumn()
@@ -46,6 +53,9 @@ export class Product {
 
   @Column({ default: false })
   is_giftcard: boolean
+
+  @DbAwareColumn({ type: "enum", enum: Status, default: "draft" })
+  status: Status
 
   @ManyToMany(() => Image, { cascade: ["insert"] })
   @JoinTable({